.clay/mcp/requirements.txt
# Python dependencies for MCP extensions

# Image processing
pillow>=10.0.0
numpy>=1.24.0

# Web requests
requests>=2.31.0

# Data processing
pandas>=2.0.0

# Natural language processing
nltk>=3.8.1
requests

google-genai